% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietcost.R
\name{addEmissionData}
\alias{addEmissionData}
\title{Emission data addition}
\usage{
addEmissionData(filepath, df, emission_cols = NULL)
}
\arguments{
\item{filepath}{Path in which the dataset, in .xlsx format, is stored.}

\item{df}{Foods dataframe.}

\item{emission_cols}{Optional parameter. Emission column names if standard dataset isn't used.}
}
\value{
Food dataframe with emission data.
}
\description{
Adds emission data to foods dataframe.
}
\examples{
foods_df <- addEmissionData(filepath = 'C:/Users/username/Downloads/dataset.xlsx', df = foods_df);
foods_df <- addEmissionData(filepath = 'C:/Users/username/Downloads/dataset.xlsx', df = foods_df, emission_cols = c('CO2','WF_l'));
}
