% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietcost.R
\name{addFoodGroupsConstraintData}
\alias{addFoodGroupsConstraintData}
\title{Food group constraint data addition}
\usage{
addFoodGroupsConstraintData(filepath, df, diets)
}
\arguments{
\item{filepath}{Path in which the dataset, in .xlsx format, is stored..}

\item{df}{Food groups dataframe.}

\item{diets}{Chosen diets. Constraint sheets in foods dataset must be of format 'constraints_DIETNAME_diet_food_groups', then the parameter passed will be DIETNAME. Can be a vector of diets in format c('DIETNAME1','DIETNAME2',...,'DIETNAMEN').}
}
\value{
Food groups dataframe with added constraint data.
}
\description{
Adds serves constraints to food groups dataframe
}
\examples{
food_groups_df <- addFoodGroupsConstraintData(filepath = 'C:/Users/username/Downloads/dataset.xlsx', food_groups_df, diets = c('C','PF', 'H'));
}
