% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNutrientData.R
\name{addNutrientData}
\alias{addNutrientData}
\title{Nutrients data addition}
\usage{
addNutrientData(filepath, df)
}
\arguments{
\item{filepath}{Path in which the dataset, in .xlsx format, is stored..}

\item{df}{Foods dataframe.}
}
\value{
Foods dataframe with nutrient columns.
}
\description{
Adds nutrients data to foods dataframe.
}
