% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietcost.R
\name{calculateResults}
\alias{calculateResults}
\title{Calculates results for a Monte Carlo Simulation}
\usage{
calculateResults(path_file, confidence_interval)
}
\arguments{
\item{path_file}{A string containing the path to the folder containing the .csv files created in the monteCarlo function.}

\item{confidence_interval}{A float. Must be either 0.01, 0.05 or 0.1.}
}
\description{
Calculates a confidence interval for several parameters obtained through a Monte Carlo Simulation. This function should be employed only if the standard table supplied with this package is utilized. Prints a .xlsx file in the home directory.
}
\examples{
calculateResults('/my/folder', 0.05)
}
