% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_function.R
\name{check_function}
\alias{check_function}
\title{Missing value check}
\usage{
check_function(name, column)
}
\arguments{
\item{name}{Column in which missing values will be sought.}

\item{column}{Column name, in string format.}
}
\description{
Checks if there are any missing values in a given column from the dataset.
}
\examples{
unique_values(df$food_id,"food ID");
}
