% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_id_defined.R
\name{check_id_defined}
\alias{check_id_defined}
\title{ID mismatch check}
\usage{
check_id_defined(df1, df2, value)
}
\arguments{
\item{df1}{First dataframe.}

\item{df2}{Second dataframe.}

\item{value}{Dataset name.}
}
\description{
Checks if a given food has an ID assigned but is absent in another dataset.
}
\examples{
check_id_defined(df1, df, 'nutrition')
}
