% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_match_individual_diet.R
\name{check_match_individual_diet}
\alias{check_match_individual_diet}
\title{Individual/diet mismatch check}
\usage{
check_match_individual_diet(df)
}
\arguments{
\item{df}{Dataframe.}
}
\value{
No return, only performs a check.
}
\description{
Checks if all individuals have a matching diet.
}
