% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_min_exists.R
\name{check_min_exists}
\alias{check_min_exists}
\title{Minimum intake food groups check}
\usage{
check_min_exists(df, check, col)
}
\arguments{
\item{df}{Dataframe.}

\item{check}{Boolean variable to permit optional food group.}

\item{col}{Minimum percentage intake column name.}
}
\value{
Dataframe.
}
\description{
Looks for a minimum value and sets zero if missing,
}
\examples{
 df <- check_min_exists(df, allow_alcohol, alcohol_perc_min)
}
