% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_nom_num_df.R
\name{check_nom_num_df}
\alias{check_nom_num_df}
\title{Applies non-nummeric value check to entire dataframe}
\usage{
check_nom_num_df(df)
}
\arguments{
\item{df}{Dataframe columns.}
}
\description{
Checks if values supposed to be numeric are in fact numeric.
}
