% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkLinkedFoods.R
\name{checkLinkedFoods}
\alias{checkLinkedFoods}
\title{Linked foods check}
\usage{
checkLinkedFoods(df, low, high)
}
\arguments{
\item{df}{Random meal plan.}

\item{low}{Vector of lower bound food IDs.}

\item{high}{Vector of higher bound food IDs.}
}
\value{
Differences dataframe.
}
\description{
Checks if lower bound linked foods serves are lower or equal to higher bound linked foods serves.
}
