% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_optional_food_groups.R
\name{checks_optional_food_groups}
\alias{checks_optional_food_groups}
\title{Optional food groups check}
\usage{
checks_optional_food_groups(check, value)
}
\arguments{
\item{check}{Boolean variable to permit optional food group.}

\item{value}{Minimum percentage of energy intake from optional food group.}
}
\value{
Minimum percentage of energy intake from optional food group.
}
\description{
If discretionary foods, alcohol or takeaway are permitted, looks for a minimum value and sets zero if missing,
}
