% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converts_dataframe.R
\name{converts_dataframe}
\alias{converts_dataframe}
\title{Weekly conversion}
\usage{
converts_dataframe(df, exclusion_cols)
}
\arguments{
\item{df}{Dataframe.}

\item{exclusion_cols}{Columns (non-numerical or percentage) that conversion won't be applied.}
}
\value{
Weekly dataframe.
}
\description{
Converts data from daily to weekly
}
\examples{
  individual_n <- grep('individual',colnames(df))
  diet_n <- grep('diet',colnames(df))
  percentage_n <- grep(percentage,colnames(df))
  exclusion_n <- c(individual_n,diet_n,percentage_n)
  df <- converts_dataframe(df, exclusion_n)
}
