% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertWeeklyFoodGroups.R
\name{convertWeeklyFoodGroups}
\alias{convertWeeklyFoodGroups}
\title{Food group serves conversion}
\usage{
convertWeeklyFoodGroups(df, diet, individual)
}
\arguments{
\item{df}{Food group serves dataframe.}

\item{diet}{Chosen diet. Must be DIETNAME from 'constraints_DIETNAME_diet_food_groups' sheet in dataset.}

\item{person}{Individual whose random meal plan will be created to. Can be one of man, woman, boy or girl.}
}
\value{
Converted food group serves dataframe.
}
\description{
Converts food group serves dataframe to weekly values.
}
\examples{
food_groups_wk <- DIETCOST::convertWeeklyFoodGroups(DIETCOST::food_groups, 'C', 'man');
}
