% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertWeeklyNutrientTargets.R
\name{convertWeeklyNutrientTargets}
\alias{convertWeeklyNutrientTargets}
\title{Nutrient targets conversion}
\usage{
convertWeeklyNutrientTargets(df, diet, person, nutrient_constraints = NULL)
}
\arguments{
\item{df}{Nutrient targets dataframe.}

\item{diet}{Chosen diet. Must be DIETNAME from 'constraints_DIETNAME_diet_foods' sheet in dataset.}

\item{person}{Individual whose random meal plan will be created to. Can be one of man, woman, boy or girl.}

\item{nutrient_constraints}{Optional parameter. Vector of nutrients column names to be used if not all nutrients are to be used as constraints.}
}
\value{
Converted nutrient targets dataframe.
}
\description{
Converts nutrient targets dataframe to weekly values.
}
\examples{
nutrient_targets_wk <- convertWeeklyNutrientTargets(nutrient_targets, 'C', 'man')
}
