% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFoodData.R
\name{createFoodData}
\alias{createFoodData}
\title{Food data creation}
\usage{
createFoodData(filepath, redmeat_ids)
}
\arguments{
\item{filepath}{Path in which the dataset, in .xlsx format, is stored.}

\item{redmeat_ids}{Vector of redmeat IDs.}
}
\value{
Food dataframe.
}
\description{
Creates a food data dataframe
}
\examples{
foods_df <- createFoodData(filepath = 'C:/Users/username/Downloads/dataset.xlsx', redmeat_ids = c(71003, 71008, 71041, 81005, 81021, 81022, 81026, 81027, 81029))
}
