% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFoodGroupData.R
\name{createFoodGroupData}
\alias{createFoodGroupData}
\title{Food group data creation}
\usage{
createFoodGroupData(df)
}
\arguments{
\item{df}{Foods dataframe.}
}
\value{
Food group dataframe.
}
\description{
Creates and populates a food group data dataframe
}
\examples{
food_groups_df <- createFoodGroupData(foods_df);
}
