% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNutrientTargets.R
\name{createNutrientTargets}
\alias{createNutrientTargets}
\title{Nutrients data addition}
\usage{
createNutrientTargets(
  filepath,
  allow_alcohol = TRUE,
  allow_discretionary = TRUE,
  allow_takeaway = TRUE,
  alcohol_perc_max = NULL,
  discretionary_perc_max = NULL,
  takeaway_perc_max = NULL
)
}
\arguments{
\item{filepath}{Path in which the dataset, in .xlsx format, is stored..}

\item{allow_alcohol}{Boolean variable checking if alcohol is permitted. Default TRUE.}

\item{allow_discretionary}{Boolean variable checking if discretionary foods are permitted. Default TRUE.}

\item{allow_takeaway}{Boolean variable checking if takeaway is permitted. Default TRUE.}

\item{alcohol_perc_max}{Optional parameter. Defines maximum energy intake derived from alcohol.}

\item{discretionary_perc_max}{Optional parameter. Defines maximum energy intake derived from discretionary foods.}

\item{takeaway_perc_max}{Optional parameter. Defines maximum energy intake derived from takeaway.}
}
\value{
Nutrient targets dataframe.
}
\description{
Adds nutrients data to foods dataframe.
}
