% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietcostExample.R
\name{dietcostExample}
\alias{dietcostExample}
\title{Package helper function}
\usage{
dietcostExample(path = NULL)
}
\arguments{
\item{path}{A string containing the path to extdata folder.}
}
\description{
Convenienc function making easy to access external data files.
}
\examples{
dietcostExample("extdata")

}
