% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietcost.R
\name{diff_calc}
\alias{diff_calc}
\title{Difference calculator}
\usage{
diff_calc(val, min, max)
}
\arguments{
\item{val}{Value to be evalueted.}

\item{min}{Minimum constraint.}

\item{max}{Maximum constraint.}
}
\value{
Difference.
}
\description{
Calculates difference between values of random meal plan created and targets logged.
}
\examples{
df$diff <- diff_calc(as.numeric(df$value), as.numeric(df$min), as.numeric(df$max));

}
