% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietcost.R
\name{energy_conversor}
\alias{energy_conversor}
\title{MJ to KJ conversion}
\usage{
energy_conversor(df, min, max)
}
\arguments{
\item{df}{Dataframe.}

\item{min}{Minimum energy column name. Default 'energy_mj_min'.}

\item{max}{Maximum energy column name. Default 'energy_mj_max'.}
}
\description{
Converts energy values in megajoules (MJ) to kilojoules (KJ),
}
\examples{
df <- energy_conversor(df, 'energy_mj_min', 'energy_mj_max')
}
