% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foodGroupData.R
\name{foodGroupData}
\alias{foodGroupData}
\title{Single-function food group dataframe creation}
\usage{
foodGroupData(filepath, df_foods, diets)
}
\arguments{
\item{filepath}{Path in which the dataset, in .xlsx format, is stored.}

\item{df_foods}{Foods dataframe.}

\item{diets}{Chosen diets. Constraint sheets in foods dataset must be of format 'constraints_DIETNAME_diet_food_groups', then the parameter passed will be DIETNAME. Can be a vector of diets in format c('DIETNAME1','DIETNAME2',...,'DIETNAMEN').}
}
\value{
Food groups dataframe.
}
\description{
Creates food groups dataframe, with constraints data, in a single function.
}
