% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietcost.R
\name{getDifference}
\alias{getDifference}
\title{General difference calculation}
\usage{
getDifference(df_target, df_nutrients, merge_col)
}
\arguments{
\item{df_target}{Constraints dataframe.}

\item{df_nutrients}{Nutrients/serves from random meal plan dataframe.}

\item{merge_col}{Column to join both dataframes.}
}
\value{
Differences dataframe.
}
\description{
Applies difference calculation to entire dataset.
}
\examples{
nutrients_diff <- getDifference(df_target = nutrient_targets_wk, df_nutrients = nutrients_plan, merge_col = 'nutrient');

}
