% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFoodGroupServes.R
\name{getFoodGroupServes}
\alias{getFoodGroupServes}
\title{Food group serves calculator}
\usage{
getFoodGroupServes(df)
}
\arguments{
\item{df}{Random meal plan.}
}
\value{
Food group serves dataframe.
}
\description{
Calculates total food group serves of random meal plan.
}
\examples{
serves <- getFoodGroupServes(random_meal);

}
