% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNutrients.R
\name{getNutrients}
\alias{getNutrients}
\title{Nutrients values calculator}
\usage{
getNutrients(df, nutrient_cols = NULL)
}
\arguments{
\item{df}{Random meal plan.}

\item{nutrient_cols}{Optional parameter. Vector of nutrients column names to be used if nutrients are different from standard dataset.}
}
\value{
Nutrients dataframe.
}
\description{
Calculates nutritional value of meal plan.
}
