% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietcost.R
\name{getPerc}
\alias{getPerc}
\title{Percentage values calculator}
\usage{
getPerc(df_nutri, df_meal)
}
\arguments{
\item{df_nutri}{Nutrient constraints dataframe.}

\item{df_meal}{Random meal plan}
}
\value{
Percentage dataframe.
}
\description{
Calculates percentage nutrient values.
}
\examples{
df <- getPerc(getNutrients(random_meal),random_meal);

}
