% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietcost.R
\name{groupedResults}
\alias{groupedResults}
\title{Aggregation of price, amount and serves results}
\usage{
groupedResults(results)
}
\arguments{
\item{results}{List of results.}
}
\value{
List of aggregated results.
}
\description{
Aggregates rsults of Monte Carlo Simulation.
}
\examples{
grouped_results <- groupedResults(results[['meal_plans']])

}
