% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_function.R
\name{join_function}
\alias{join_function}
\title{Join function}
\usage{
join_function(df1, df2, condition)
}
\arguments{
\item{df1}{First dataframe.}

\item{df2}{Second dataframe.}

\item{condition}{Column in which the two datframes will be joined. Can be a single string or a vector.}
}
\value{
Dataframe.
}
\description{
Safely performs a left join between two dataframes.
}
