% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrientDataCalculation.R
\name{nutrientDataCalculation}
\alias{nutrientDataCalculation}
\title{Nutrient data application to random meal plan created}
\usage{
nutrientDataCalculation(df, nutrient_cols = NULL)
}
\arguments{
\item{df}{Random meal plan.}

\item{nutrient_cols}{Optional parameter. Nutrient column names if standard dataset isn't used.}
}
\value{
Random meal plan with nutrients calculated.
}
\description{
Applies nutrient data calculation to random meal plan generated.
}
