% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permitted_individuals.R
\name{permitted_individuals}
\alias{permitted_individuals}
\title{Permitted individuals check}
\usage{
permitted_individuals(df)
}
\arguments{
\item{df}{Variable.}
}
\description{
Checks if logged individuals are one or all of the following: man, woman, boy or girl.
}
\examples{
permitted_individuals(data.frame(individual = person))
}
