% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietcost.R
\name{priceEmissionData}
\alias{priceEmissionData}
\title{Price/emission data application to random meal plan created}
\usage{
priceEmissionData(df, emission_cols = NULL)
}
\arguments{
\item{df}{Random meal plan.}

\item{emission_cols}{Optional parameter. Emission column names if standard dataset isn't used.}
}
\value{
Random meal plan with price and emissions calculated.
}
\description{
Applies price and emission data calculation to random meal plan generated.
}
\examples{
df <-priceEmissionData(meal_plan);
df <-priceEmissionData(meal_plan, c('CO2','WF_l'));
}
