% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printResults.R
\name{printResults}
\alias{printResults}
\title{Exportation of Monte Carlo results}
\usage{
printResults(results, person, diet, allowed_varieties, iterations)
}
\arguments{
\item{results}{List of results}

\item{person}{Individual whose random meal plan will be created to. Can be one of man, woman, boy or girl.}

\item{diet}{Chosen diet. Must be DIETNAME from 'constraints_DIETNAME_diet_foods' sheet in dataset.}

\item{allowed_varieties}{Permitted food varieties. Can be a vector of the following: 1,2 and/or 3.}

\item{iterations}{Number of iterations. Integer.}
}
\description{
Exports, in .xlsx format, the results of Monte Carlo simulation.
}
\examples{
printResults(results, person, diet, allowed_varieties, iterations)#' 
}
