% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_plan.R
\name{random_plan}
\alias{random_plan}
\title{Random deletion}
\usage{
random_plan(df, column, condition)
}
\arguments{
\item{df}{Dataframe.}

\item{column}{Column from which decision about removal of values will be made.}

\item{condition}{Condition that, if is true, will enable radom removal.}
}
\value{
Random meal plan
}
\description{
Randomly deletes a food.
}
