% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietcost.R
\name{redmeat_check}
\alias{redmeat_check}
\title{Redmeat flag}
\usage{
redmeat_check(id, redmeat_ids)
}
\arguments{
\item{id}{Food group id column in dataframe.}

\item{redmeat_ids}{Vector of unique food IDs that are redmeat.}
}
\description{
Sets a boolean redmeat flag column in dataset.
}
\examples{
df$redmeat <- sapply(df$food_id, redmeat_check, c(00001, 00002))
}
