% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietcost.R
\name{remove_suffix}
\alias{remove_suffix}
\title{Suffix removal}
\usage{
remove_suffix(vector, suffix_1, suffix_2)
}
\arguments{
\item{vector}{Vector of column names}

\item{suffix_1}{First suffix to be removed.}

\item{suffix_2}{Second suffix to be removed.}
}
\value{
Vector of column names without suffixes.
}
\description{
Removes one of two suffixes from column names
}
\examples{
nutrient_names <- remove_suffix(nutrient_names, '_grams','_mgrams')

}
