% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_safe.R
\name{sample_safe}
\alias{sample_safe}
\title{Safe sampling}
\usage{
sample_safe(x)
}
\arguments{
\item{x}{Vector.}
}
\value{
Random sample.
}
\description{
Safely extracts a random unitary sample from a vector.
}
\examples{
serve_range <- c(10,25,37,52,100);
foods_df$intake <- sample_safe(serve_range);
}
