% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sauces_protein_discretionary_change.R
\name{sauces_protein_discretionary_change}
\alias{sauces_protein_discretionary_change}
\title{Sauces, protein and discretionary food groups treatment}
\usage{
sauces_protein_discretionary_change(group)
}
\arguments{
\item{group}{Food group column in dataframe.}
}
\value{
Treated dataframe.
}
\description{
Treats above said food name groups to the format used in the package.
}
\examples{
df$food_group <- sapply(df$food_group, sauces_protein_discretionary_change)
}
