% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starchy_fill.R
\name{starchy_fill}
\alias{starchy_fill}
\title{Starchy vegetables serves addition}
\usage{
starchy_fill(df, starchy_name, serve_identifier, max_identifier)
}
\arguments{
\item{df}{Dataframe.}

\item{starchy_name}{Starchy vegetables food group name. Default 'Starchy vegetables'.}

\item{serve_identifier}{Serve column identifier. Default 'serve'.}

\item{max_identifier}{Max column identifier. Default 'max'.}
}
\value{
Food group dataframe with starchy vegetable minimum and maximum serves columns added.
}
\description{
Adds minimum and maximum serves of starchy vegetables.
}
