% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treat_df.R
\name{treat_df}
\alias{treat_df}
\title{Pre-treatment of constraint data}
\usage{
treat_df(df, min_identifier, max_identifier, suffix, max_scale, override_min)
}
\arguments{
\item{df}{Dataframe to be treated.}

\item{min_identifier}{Minimum value column identifier. 'Min' in standard dataset.}

\item{max_identifier}{Maximum value column identifier. 'Max' in standard dataset.}

\item{suffix}{Suffix to be added to column name.}

\item{max_scale}{Maximum scale. Default is two.}

\item{override_min}{If is not null, overrides all minimum values.}
}
\value{
Treated dataframe.
}
\description{
Pre-treatment of constraints dataframe.
}
\examples{
treat_df(df1,min, max, diet_suffix, max_scale, override_min)
}
