% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treat_groups_df.R
\name{treat_groups_df}
\alias{treat_groups_df}
\title{Treatment of food group constraints dataframe}
\usage{
treat_groups_df(df, suffix)
}
\arguments{
\item{df}{Dataframe.}

\item{suffix}{Suffix to be added to column.}
}
\value{
Treated food group dataframe.
}
\description{
Converts weekly food group serves to daily and adds diet suffix to column names.
}
\examples{
df <- treat_groups_df(df, 'C')
}
