% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_values.R
\name{unique_values}
\alias{unique_values}
\title{Unique value check}
\usage{
unique_values(value, df, value_col, value_name)
}
\arguments{
\item{value}{Column from which an unique vector will be formed.}

\item{df}{Dataframe in which lies the column to be checked.}

\item{value_col}{Name of the column to be checked, in string format.}

\item{value_name}{Name of the variable tested.}
}
\description{
Checks if there are non-unique values in dataset.
}
