% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_data.R
\name{upload_data}
\alias{upload_data}
\title{Data upload}
\usage{
upload_data(filepath, sheet)
}
\arguments{
\item{filepath}{The filepath in which the dataset, in .xlsx format, is saved.}

\item{sheet}{The sheet of the .xlsx to be read.}
}
\value{
The dataframe generated by the file which was read.
}
\description{
Safely uploads data to be processed in DIETCOST software.
}
