% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPriceData.R
\name{addPriceData}
\alias{addPriceData}
\title{Price data addition}
\usage{
addPriceData(filepath, df)
}
\arguments{
\item{filepath}{Path in which the dataset, in .xlsx format, is stored..}

\item{df}{Foods dataframe.}
}
\value{
Foods dataframe with added price data.
}
\description{
Adds price data to foods dataframe.
}
