% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_match_food_price.R
\name{check_match_food_price}
\alias{check_match_food_price}
\title{Food/price mismatch check}
\usage{
check_match_food_price(df)
}
\arguments{
\item{df}{Dataframe.}
}
\description{
Checks if all foods have a price.
}
