% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_non_num.R
\name{check_non_num}
\alias{check_non_num}
\title{Non-numeric check}
\usage{
check_non_num(df)
}
\arguments{
\item{df}{Dataframe column.}
}
\description{
Checks if values supposed to be numeric are in fact numeric.
}
\examples{
check_nom_num(df$food_id)
}
