% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_spelling.R
\name{check_spelling}
\alias{check_spelling}
\title{Spellcheck}
\usage{
check_spelling(df1, df2, condition)
}
\arguments{
\item{df1}{First dataframe.}

\item{df2}{Second dataframe.}

\item{condition}{Column to be joined.}
}
\value{
No return, only performs a check.
}
\description{
Checks if two datasets have the same spelling in names column.
}
