% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_variety.R
\name{check_variety}
\alias{check_variety}
\title{Variety check}
\usage{
check_variety(df)
}
\arguments{
\item{df}{Dataframe variety column.}
}
\description{
Checks if varieties are into the allowed range (1,2 or 3).
}
\examples{
DIETCOST::check_variety(DIETCOST::foods$variety)
}
