% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkLinkedFoods.R
\name{checkLinkedFoods}
\alias{checkLinkedFoods}
\title{Linked foods check}
\usage{
checkLinkedFoods(df, low, high)
}
\arguments{
\item{df}{Random meal plan.}

\item{low}{Vector of lower bound food IDs.}

\item{high}{Vector of higher bound food IDs.}
}
\value{
Differences dataframe.
}
\description{
Checks if lower bound linked foods serves are lower or equal to higher bound linked foods serves.
}
\examples{
linked_milk_low <- c("69029", "79012", "70038", "69043");
linked_milk_high <- c("79020", "79001", "79036", "79003", "79039");
link_sum_1 <- checkLinkedFoods(random_meal, linked_milk_low, linked_milk_high);

}
