% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietcost.R
\name{checkZeroDiff}
\alias{checkZeroDiff}
\title{All zero difference check}
\usage{
checkZeroDiff(diff)
}
\arguments{
\item{diff}{Differences dataframe}
}
\value{
Boolean. TRUE if all zeroes, FALSE otherwise.
}
\description{
Checks if differences dataframe is all zeroes.
}
\examples{
checkZeroDiff(nutrients_diff)

}
