% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dietcost.R
\name{random_plan}
\alias{random_plan}
\title{Random deletion}
\usage{
random_plan(df, column, condition)
}
\arguments{
\item{df}{Dataframe.}

\item{column}{Condition that, if is true, will enable radom removal.}
}
\value{
Random meal plan
}
\description{
Randomly deletes a food.
}
\examples{
foods_df <- random_plan(foods_df, 'food_group_id', discretionary_id)
}
