% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_name_check.R
\name{standard_name_check}
\alias{standard_name_check}
\title{Standard name check}
\usage{
standard_name_check(df, ...)
}
\arguments{
\item{df}{Dataframe.}

\item{...}{Any number of strings.}
}
\description{
Checks if variable names are the standard defined into DIETCOST R standard table.
}
\examples{
standard_name_check(foods, 'food_id', 'food_name')
}
